feat: move tab selection block with >> / <<
When multiple tabs are highlighted (via zz/ZZ), pressing >> or << now moves
the entire selection block together as a unit. Moves the adjacent outside tab
to the opposite side of the block — one swap per step, count supported.

Extract all selection logic (selectNextTabForGroup, selectPreviousTabForGroup,
moveTabSelection) into a new tab_selection.js module. tab_groups.js stays
focused on group-aware single-tab movement and delegates to moveTabSelection
when multiple non-pinned tabs are highlighted.
