+++ tests/unit_tests/tab_groups_test.js
@@ -1,20 +1,6 @@
 import "./test_helper.js";
 import "../../background_scripts/bg_utils.js";
-import {
-  moveTab,
-  selectNextTabForGroup,
-  selectPreviousTabForGroup,
-} from "../../background_scripts/tab_groups.js";
-
-function makeTabs(count, highlighted = []) {
-  return Array.from({ length: count }, (_, i) => ({
-    id: i + 1,
-    index: i,
-    windowId: 1,
-    highlighted: highlighted.includes(i),
-    groupId: -1,
-  }));
-}
+import { moveTab } from "../../background_scripts/tab_groups.js";
 
 context("moveTab (>> / <<) past a collapsed group at the window edge", () => {
   let movedTo;
@@ -50,118 +36,3 @@ context("moveTab (>> / <<) past a collapsed group at the window edge", () => {
   });
 });
 
-context("selectNextTabForGroup (zz)", () => {
-  let capturedHighlight;
-
-  setup(() => {
-    capturedHighlight = null;
-    stub(chrome.tabs, "highlight", (args) => {
-      capturedHighlight = args.tabs;
-    });
-  });
-
-  should("extend right when only the anchor is selected", async () => {
-    const tabs = makeTabs(5, [2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectNextTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2, 3], capturedHighlight);
-  });
-
-  should("extend further right when right side is already extended", async () => {
-    const tabs = makeTabs(5, [2, 3]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectNextTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2, 3, 4], capturedHighlight);
-  });
-
-  should("shrink from left when left side is extended", async () => {
-    const tabs = makeTabs(5, [1, 2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectNextTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2], capturedHighlight);
-  });
-
-  should("do nothing when anchor is at the last tab", async () => {
-    const tabs = makeTabs(3, [2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectNextTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal(null, capturedHighlight);
-  });
-
-  should("do nothing when right extension already reaches the last tab", async () => {
-    const tabs = makeTabs(3, [1, 2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectNextTabForGroup({ tab: tabs[1], count: 1 });
-    assert.equal(null, capturedHighlight);
-  });
-
-  should("extend right 3 tabs with count: 3", async () => {
-    const tabs = makeTabs(6, [2]);
-    let callCount = 0;
-    stub(chrome.tabs, "query", () => {
-      const highlighted = [2, ...Array.from({ length: callCount }, (_, i) => 3 + i)];
-      callCount++;
-      return makeTabs(6, highlighted);
-    });
-    await selectNextTabForGroup({ tab: tabs[2], count: 3 });
-    assert.equal([2, 3, 4, 5], capturedHighlight);
-  });
-});
-
-context("selectPreviousTabForGroup (ZZ)", () => {
-  let capturedHighlight;
-
-  setup(() => {
-    capturedHighlight = null;
-    stub(chrome.tabs, "highlight", (args) => {
-      capturedHighlight = args.tabs;
-    });
-  });
-
-  should("extend left when only the anchor is selected", async () => {
-    const tabs = makeTabs(5, [2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectPreviousTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2, 1], capturedHighlight);
-  });
-
-  should("extend further left when left side is already extended", async () => {
-    const tabs = makeTabs(5, [1, 2]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectPreviousTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2, 1, 0], capturedHighlight);
-  });
-
-  should("shrink from right when right side is extended", async () => {
-    const tabs = makeTabs(5, [2, 3]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectPreviousTabForGroup({ tab: tabs[2], count: 1 });
-    assert.equal([2], capturedHighlight);
-  });
-
-  should("do nothing when anchor is at the first tab", async () => {
-    const tabs = makeTabs(3, [0]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectPreviousTabForGroup({ tab: tabs[0], count: 1 });
-    assert.equal(null, capturedHighlight);
-  });
-
-  should("do nothing when left extension already reaches the first tab", async () => {
-    const tabs = makeTabs(3, [0, 1]);
-    stub(chrome.tabs, "query", () => tabs);
-    await selectPreviousTabForGroup({ tab: tabs[1], count: 1 });
-    assert.equal(null, capturedHighlight);
-  });
-
-  should("extend left 3 tabs with count: 3", async () => {
-    const tabs = makeTabs(6, [3]);
-    let callCount = 0;
-    stub(chrome.tabs, "query", () => {
-      const highlighted = [3, ...Array.from({ length: callCount }, (_, i) => 2 - i)];
-      callCount++;
-      return makeTabs(6, highlighted);
-    });
-    await selectPreviousTabForGroup({ tab: tabs[3], count: 3 });
-    assert.equal([3, 1, 2, 0], capturedHighlight);
-  });
-});
